test: expand coverage with algebra tests
Added test coverage for coreason_manifest.utils.algebra.py including `verify_ast_safety`, `verify_merkle_proof`, and `apply_state_differential`. Removed the unauthorized `tests/test_new_ontology_validators.py` and placed the relevant logic properly into `tests/contracts/test_algebra.py`.
